test(tree): move state root task test to payload processor
