+++ crates/engine/tree/Cargo.toml
@@ -70,10 +70,11 @@ reth-tracing = { workspace = true, optional = true }
 # reth
 reth-chain-state = { workspace = true, features = ["test-utils"] }
 reth-chainspec.workspace = true
-reth-ethereum-engine-primitives.workspace = true
+reth-db-common.workspace = true
 reth-ethereum-consensus.workspace = true
-reth-evm-ethereum.workspace = true
+reth-ethereum-engine-primitives.workspace = true
 reth-evm = { workspace = true, features = ["test-utils"] }
+reth-evm-ethereum.workspace = true
 reth-exex-types.workspace = true
 reth-network-p2p = { workspace = true, features = ["test-utils"] }
 reth-prune-types.workspace = true
